requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=0.6,!=0.7,<1.0
Babel>=1.3
Flask>=0.10,<1.0
iso8601>=0.1.9
oslo.config>=1.6.0          # Apache-2.0
oslo.i18n>=1.3.0            # Apache-2.0
oslo.log>=0.4.0             # Apache-2.0
oslo.serialization>=1.2.0   # Apache-2.0
oslo.utils>=1.2.0           # Apache-2.0
paramiko>=1.13.0
psutil>=1.1.1,<2.0.0
PyGithub
python-memcached>=1.48
PyYAML>=3.1.0
sh
six>=1.9.0
